refactor(backend): refactor backend code structure
